feat: apply react router for pages in project
