--- src/ContactEditPage/index.tsx
@@ -1,63 +1,0 @@
-import { Form, useLoaderData, redirect } from 'react-router-dom';
-import { Contact, updateContact } from 'src/contacts';
-
-export async function action({ request, params }) {
-  const formData = await request.formData();
-  const updates = Object.fromEntries(formData);
-  await updateContact(params.contactId, updates);
-  return redirect(`/contacts/${params.contactId}`);
-}
-
-export function ContactEditPage() {
-  const { contact } = useLoaderData() as Contact;
-  return (
-    <Form method="post" id="contact-form">
-      <p>
-        <span>Name</span>
-        <input
-          placeholder="First"
-          aria-label="First name"
-          type="text"
-          name="first"
-          defaultValue={contact.first}
-        />
-        <input
-          placeholder="Last"
-          aria-label="Last name"
-          type="text"
-          name="last"
-          defaultValue={contact.last}
-        />
-      </p>
-      <label htmlFor="twitter">
-        <span>Twitter</span>
-        <input
-          id="twitter"
-          type="text"
-          name="twitter"
-          placeholder="@jack"
-          defaultValue={contact.twitter}
-        />
-      </label>
-      <label htmlFor="avatar">
-        <span>Avatar URL</span>
-        <input
-          id="avatar"
-          placeholder="https://example.com/avatar.jpg"
-          aria-label="Avatar URL"
-          type="text"
-          name="avatar"
-          defaultValue={contact.avatar}
-        />
-      </label>
-      <label htmlFor="notes">
-        <span>Notes</span>
-        <textarea id="notes" name="notes" defaultValue={contact.notes} rows={6} />
-      </label>
-      <p>
-        <button type="submit">Save</button>
-        <button type="button">Cancel</button>
-      </p>
-    </Form>
-  );
-}
