+++ src/common/pages/Home/index.tsx
@@ -1,0 +1,3 @@
+export function Home() {
+  return <p>Homepage</p>;
+}
