+++ src/demo/lazy-load-image/demo-2.tsx
