--- src/demo/lazy-load-page/Page1.tsx
@@ -1,3 +1,0 @@
-export function Page1() {
-  return <p>content 1</p>;
-}
