--- src/demo/lazy-load-page/Page2.tsx
@@ -1,10 +1,0 @@
-export function Page2() {
-  const data = 'I came from the About.tsx loader function!';
-
-  return (
-    <div>
-      <h2>About</h2>
-      <p>{data}</p>
-    </div>
-  );
-}
