--- src/demo/multi-app/customer/AccountEdit/index.tsx
@@ -1,12 +1,0 @@
-import { Link } from 'react-router-dom';
-
-export function AccountEdit() {
-  return (
-    <>
-      <p>account edit</p>
-      <p>
-        <Link to="notfound">Go to not found page</Link>
-      </p>
-    </>
-  );
-}
