--- src/demo/multi-app/customer/AccountInfo/index.tsx
@@ -1,9 +1,0 @@
-import { Link } from 'react-router-dom';
-
-export function AccountInfo() {
-  return (
-    <p>
-      account info. <Link to="edit">Edit</Link>
-    </p>
-  );
-}
