--- src/demo/multi-app/sale/OrderDetail/index.tsx
@@ -1,6 +1,0 @@
-import { useParams } from 'react-router-dom';
-
-export function OrderDetail() {
-  const { id } = useParams();
-  return <p>order detail {id}</p>;
-}
